chore: refactor run service
