+++ run-service/server.js
@@ -1,9 +1,9 @@
-import express from 'express';
-import { createClient } from 'redis';
-import dotenv from 'dotenv';
-import router from './router.js';
-import corsMiddleware from './middlewares/cors.js';
-import loggingMiddleware from './middlewares/logging.js';
+import express from "express";
+import { createClient } from "redis";
+import dotenv from "dotenv";
+import router from "./router/router.js";
+import corsMiddleware from "./middlewares/cors.js";
+import loggingMiddleware from "./middlewares/logging.js";
 import errorHandler from "./middlewares/errorHandler.js";
 
 const app = express();
@@ -15,23 +15,23 @@ const questionServiceUrl = process.env.QUESTION_SVC_DEV;
 
 // TODO: Replace with actual Redis connection instance
 const redisClient = createClient({
-  url: process.env.REDIS_URL_DEV ?? 'redis://localhost:6379',
+  url: process.env.REDIS_URL_DEV ?? "redis://localhost:6379",
 });
 
-redisClient.on('error', (err) => console.log('Redis Client Error', err));
-redisClient.on('connect', () => console.log('Connected to Redis'));
-redisClient.on('ready', () => console.log('Redis Client Ready'));
+redisClient.on("error", (err) => console.log("Redis Client Error", err));
+redisClient.on("connect", () => console.log("Connected to Redis"));
+redisClient.on("ready", () => console.log("Redis Client Ready"));
 await redisClient.connect();
 
 app.use(loggingMiddleware);
 app.use(corsMiddleware);
-app.use('/api/run-service', router);
+app.use("/api/run-service", router);
 
 // Test Route for Health Checks
-app.get('/healthz', (req, res) => {
+app.get("/healthz", (req, res) => {
   res
     .status(200)
-    .json({ message: 'Connected to /healthz route of run-service' });
+    .json({ message: "Connected to /healthz route of run-service" });
 });
 
 app.use(errorHandler);
